Rename frontend_wayland protocol files to their canonical names
Six protocol implementations were named after neither the protocol XML
nor the interface they implement. Rename them to match the protocol:

  fractional_scale_v1            -> wp_fractional_scale_v1
  pointer_constraints_unstable_v1 -> zwp_pointer_constraints_v1
  relative_pointer_unstable_v1    -> zwp_relative_pointer_v1
  xdg_activation_v1               -> xdg_activation_unstable_v1
  xdg_decoration_unstable_v1      -> xdg_decoration_manager_v1
  xdg_dialog_v1                   -> xdg_wm_dialog_v1

Pure file renames plus the corresponding #include, header guard and
CMakeLists updates. No behaviour change; the factory function and class
names are untouched.

(xdg_decoration_manager_v1.h also had a copy-pasted closing guard
comment naming the relative-pointer guard; corrected here.)
