+++ src/server/frontend_wayland/CMakeLists.txt
@@ -15,8 +15,8 @@ set(
   wl_data_device.cpp            wl_data_device.h
   wl_data_source.cpp            wl_data_source.h
   output_manager.cpp            output_manager.h
-  pointer_constraints_unstable_v1.cpp pointer_constraints_unstable_v1.h
-  relative_pointer_unstable_v1.cpp    relative_pointer_unstable_v1.h
+  zwp_pointer_constraints_v1.cpp     zwp_pointer_constraints_v1.h
+  zwp_relative_pointer_v1.cpp        zwp_relative_pointer_v1.h
   wl_subcompositor.cpp          wl_subcompositor.h
                                 wl_surface_role.h
   window_wl_surface_role.cpp    window_wl_surface_role.h
@@ -53,8 +53,8 @@ set(
   text_input_v1.cpp             text_input_v1.h
   primary_selection_v1.cpp      primary_selection_v1.h
   session_lock_v1.cpp           session_lock_v1.h
-  xdg_decoration_unstable_v1.cpp           xdg_decoration_unstable_v1.h
-  xdg_dialog_v1.cpp                        xdg_dialog_v1.h
+  xdg_decoration_manager_v1.cpp            xdg_decoration_manager_v1.h
+  xdg_wm_dialog_v1.cpp                     xdg_wm_dialog_v1.h
   server_decoration_manager.cpp            server_decoration_manager.h
   ${PROJECT_SOURCE_DIR}/src/include/server/mir/frontend/wayland.h
   ${CMAKE_CURRENT_BINARY_DIR}/wayland_frontend.tp.c
@@ -65,8 +65,8 @@ set(
   session_credentials.cpp
   ${PROJECT_SOURCE_DIR}/src/include/server/mir/frontend/buffer_stream.h
   wp_viewporter.cpp             wp_viewporter.h
-  fractional_scale_v1.cpp           fractional_scale_v1.h
-  xdg_activation_v1.cpp         xdg_activation_v1.h
+  wp_fractional_scale_v1.cpp        wp_fractional_scale_v1.h
+  xdg_activation_unstable_v1.cpp xdg_activation_unstable_v1.h
   linux_drm_syncobj.cpp         linux_drm_syncobj.h
   data_control_v1.cpp           data_control_v1.h
   surface_registry.cpp          surface_registry.h
